[G] feat: match winning numbers

+++ src/main/java/lotto/prize/MatchResult.java
@@ -9,16 +9,16 @@ import java.util.Objects;
 public class MatchResult {
 
     private final Money payment;
-    private final Map<LottoPrize, Integer> matchResult;
+    private final Map<LottoPrize, Long> matchResult;
 
-    public MatchResult(Money payment, Map<LottoPrize, Integer> matchResult) {
+    public MatchResult(Money payment, Map<LottoPrize, Long> matchResult) {
         validate(payment, matchResult);
 
         this.payment = payment;
         this.matchResult = matchResult;
     }
 
-    private void validate(Money payment, Map<LottoPrize, Integer> matchResult) {
+    private void validate(Money payment, Map<LottoPrize, Long> matchResult) {
         if (Objects.isNull(payment)) {
             throw new IllegalArgumentException("Payment can't be null");
         }
@@ -28,7 +28,7 @@ public class MatchResult {
         }
     }
 
-    public static MatchResult init(Money payment, Map<LottoPrize, Integer> matchResult) {
+    public static MatchResult init(Money payment, Map<LottoPrize, Long> matchResult) {
         return new MatchResult(payment, matchResult);
     }
 
@@ -46,7 +46,8 @@ public class MatchResult {
     }
 
     public int matchCount(LottoPrize lottoPrize) {
-        return matchResult.get(lottoPrize);
+        return matchResult.get(lottoPrize)
+                .intValue();
     }
 
 }
